refact: relocate service
