--- app/dashboard/src/components/modules/Router.tsx
@@ -1,10 +1,0 @@
-import { fetch } from '../../service/http';
-import { getAuthToken } from '../../utils/authStorage';
-
-export const fetchAdminLoader = () => {
-  return fetch('/admin', {
-    headers: {
-      Authorization: `Bearer ${getAuthToken()}`,
-    },
-  });
-};
